requirements.txt
atomicwrites==1.4.0
attrs==20.1.0
click==7.1.2
colorama==0.4.3
coverage==5.2.1
flake8==3.8.3
Flask==1.1.2
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
picamera==1.8
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pynput==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
six==1.15.0
tinkerforge==2.1.26
toml==0.10.1
Werkzeug==1.0.1
